Feat/spring-ai (#29) deepseek provider
* refactor: remove ping method and related tests from AiApiClientInterface

* feat: add DeepSeekAdapter and integrate spring-ai dependency

* feat: add spring-ai library and create DeepSeekAdapter with message merging logic

* refactor: refactor adapters and controllers to use Lombok's @AllArgsConstructor for cleaner code

* fix: lower minimum coverage thresholds to 0.91 for line and instruction metrics

---------

+++ src/main/java/com/delimce/aibroker/domain/dto/requests/llm/ModelRequest.java
@@ -28,5 +28,5 @@ public class ModelRequest {
     ModelMessageRequest[] messages;
 
     @Builder.Default
-    int temperature = 1;
+    double temperature = 1.0;
 }
